refactor(ui): rehome deterministic id
