+++ npm/ui/core/src/families/form/search-field/search-field.vue
@@ -2,7 +2,10 @@
 import { computed, nextTick, ref, useTemplateRef, watch, watchEffect } from "vue";
 
 import { useControllableState } from "../../../controllable-state.ts";
-import { deriveDeterministicId, useDeterministicId } from "../../../deterministic-id.ts";
+import {
+  deriveDeterministicId,
+  useDeterministicId,
+} from "../../foundations/id/deterministic-id.ts";
 import type {
   SearchFieldAriaInvalid,
   SearchFieldClearSlotState,
