+++ npm/ui/core/src/families/navigation/pagination/pagination.vue
@@ -3,7 +3,10 @@ import { computed, useTemplateRef } from "vue";
 import type { ComputedRef } from "vue";
 
 import { useControllableState } from "../../../controllable-state.ts";
-import { deriveDeterministicId, useDeterministicId } from "../../../deterministic-id.ts";
+import {
+  deriveDeterministicId,
+  useDeterministicId,
+} from "../../foundations/id/deterministic-id.ts";
 import { paginationContext } from "./pagination-context.ts";
 import {
   createPaginationRange,
