+++ npm/ui/core/src/families/overlays/dialog/dialog-root.vue
@@ -3,7 +3,10 @@ import { computed, shallowRef } from "vue";
 import type { ComputedRef } from "vue";
 
 import { useControllableState } from "../../../controllable-state.ts";
-import { deriveDeterministicId, useDeterministicId } from "../../../deterministic-id.ts";
+import {
+  deriveDeterministicId,
+  useDeterministicId,
+} from "../../foundations/id/deterministic-id.ts";
 import { dialogContext } from "./dialog-context.ts";
 import type { DialogRootExpose, DialogSlotState, DialogState } from "./dialog-types.ts";
