+++ npm/ui/core/src/families/selection/radio-group/radio-group-item.vue
@@ -2,7 +2,7 @@
 import { computed, nextTick, useTemplateRef, watchEffect } from "vue";
 import type { ComputedRef } from "vue";
 
-import { useDeterministicId } from "../../../deterministic-id.ts";
+import { useDeterministicId } from "../../foundations/id/deterministic-id.ts";
 import { radioGroupContext } from "./radio-group-context.ts";
 import type { RadioGroupItemExpose, RadioGroupItemState } from "./radio-group-types.ts";
