+++ npm/ui/core/src/families/selection/switch/switch-control.vue
@@ -2,7 +2,7 @@
 import { computed, useTemplateRef, watch } from "vue";
 
 import { useControllableState } from "../../../controllable-state.ts";
-import { useDeterministicId } from "../../../deterministic-id.ts";
+import { useDeterministicId } from "../../foundations/id/deterministic-id.ts";
 import type {
   SwitchAriaInvalid,
   SwitchExpose,
