+++ npm/ui/core/src/sfc-lint.test.ts
@@ -22,7 +22,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/collapsible-content.vue",
       "src/collapsible-root.vue",
       "src/collapsible-trigger.vue",
-      "src/deterministic-id-provider.vue",
       "src/error-summary.vue",
       "src/families/accessibility/announcer/announcer-provider.vue",
       "src/families/accessibility/live-region/live-region.vue",
@@ -64,6 +63,7 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/families/form/search-field/search-field.vue",
       "src/families/form/slider/slider.vue",
       "src/families/form/textarea/textarea-control.vue",
+      "src/families/foundations/id/deterministic-id-provider.vue",
       "src/families/i18n/locale/locale-provider.vue",
       "src/families/layout/aspect-ratio/aspect-ratio.vue",
       "src/families/layout/avatar/avatar.vue",
@@ -147,7 +147,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "collapsible-content.vue",
       "collapsible-root.vue",
       "collapsible-trigger.vue",
-      "deterministic-id-provider.vue",
       "error-summary.vue",
       "families/accessibility/announcer/announcer-provider.vue",
       "families/accessibility/live-region/live-region.vue",
@@ -189,6 +188,7 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "families/form/search-field/search-field.vue",
       "families/form/slider/slider.vue",
       "families/form/textarea/textarea-control.vue",
+      "families/foundations/id/deterministic-id-provider.vue",
       "families/i18n/locale/locale-provider.vue",
       "families/layout/aspect-ratio/aspect-ratio.vue",
       "families/layout/avatar/avatar.vue",
